Getting ready for Euro 2024
Updated Ruby interpreter to 3.4 & all gems to their most current
version.

Also added a script for the raw data for the Euro fixture list. I
foolishly put this into the test data, which broke a lot of the tests.
On the test data, I made it more reliable by fixing all the dates to be
relative to current date and time.

There's also a lot of whitespace changes in the tests as Rubocop has had
his wiked way with the tests as well as the source code.

Think that's about it. Am committing with all tests passing but still
need to make functional changes to the website such as accounting for
the fewer groups